Adding a test for a section
