+++ openquake/hazardlib/tests/geo/surface/data/section_1680_ucf.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<nrml
+xmlns="http://openquake.org/xmlns/nrml/0.5"
+xmlns:gml="http://www.opengis.net/gml"
+>
+    <geometryModel
+    name="sections_rupture200"
+    >
+                <section
+        id="1680"
+        >
+            <kiteSurface>
+                <profile>
+                    <gml:LineString>
+                        <gml:posList>
+                            -122.0011 37.6073 0.0 -122.0011 37.6072 5.1
+                        </gml:posList>
+                    </gml:LineString>
+                </profile>
+                <profile>
+                    <gml:LineString>
+                        <gml:posList>
+                            -121.9813 37.5952 0.0 -121.9813 37.5952 5.1
+                        </gml:posList>
+                    </gml:LineString>
+                </profile>
+                <profile>
+                    <gml:LineString>
+                        <gml:posList>
+                            -121.9802 37.5943 0.0 -121.9803 37.5942 5.1
+                        </gml:posList>
+                    </gml:LineString>
+                </profile>
+            </kiteSurface>
+        </section>
+    </geometryModel>
+</nrml>
